requirements.txt
requests
beautifulsoup4
pdfplumber
pandas
mysql-connector-python
